Make normalize_includes public
This function is useful for converting included params from the client
into Ecto's `preload/3` function.
